requirements.txt
numpy==1.16.2
Pillow==5.4.1
PyQt5==5.13.1
PyQt5-stubs==5.13.1.2